initial

+++ app/src/main/java/com/mobile/agri10x/models/GetOrderListDatumBookingCartData.java
@@ -15,7 +15,7 @@ public class GetOrderListDatumBookingCartData {
         private List<GetOrderListDatumBookingCartDataProduct> products = null;
         @SerializedName("subTotal")
         @Expose
-        private Integer subTotal;
+        private Double subTotal;
 
         public List<GetOrderListDatumBookingCartDataProduct> getProducts() {
             return products;
@@ -25,11 +25,11 @@ public class GetOrderListDatumBookingCartData {
             this.products = products;
         }
 
-        public Integer getSubTotal() {
+        public double getSubTotal() {
             return subTotal;
         }
 
-        public void setSubTotal(Integer subTotal) {
+        public void setSubTotal(double subTotal) {
             this.subTotal = subTotal;
         }
